pyproject.toml
[project]
name = "chess-tui"
version = "0.1.0"
description = "Terminal-native chess study workstation — multi-engine analysis, opening explorer, puzzle trainer"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.80,<10",
    "python-chess>=1.10,<2",
]

[project.optional-dependencies]
agent = ["aiohttp>=3.8"]
advisor = ["anthropic>=0.30"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["chess_tui"]

[tool.setuptools.package-data]
chess_tui = ["tui.tcss"]
